requirements.txt
PyJWT==1.7.1
cryptography==2.8
Flask==1.1.1
flask-marshmallow==0.11.0
flask-restplus==0.13.0
marshmallow==3.4.0
pymongo==3.9.0
Werkzeug==0.16.1
pytest==5.3.5
docker==4.2.0
PyYAML==5.3
ansible==2.9.5